+++ mutwo/core_utilities/exceptions.py
@@ -1,40 +1,83 @@
 """Module for mutwo specific exceptions."""
 
-import typing
-
 from mutwo import core_constants
 
 __all__ = (
-    "ValueNotAssignedError",
-    "OverlappingTimeBracketsError",
+    "AlreadyDefinedValueNameError",
+    "InvalidAverageValueStartAndEndWarning",
+    "InvalidStartValueError",
+    "InvalidPointError",
+    "ImpossibleToSquashInError",
+    "InvalidStartAndEndValueError",
+    "InvalidCutOutStartAndEndValuesError",
     "SplitUnavailableChildError",
     "NoSolutionFoundError",
 )
 
 
-class ValueNotAssignedError(Exception):
-    def __init__(self):
+class AlreadyDefinedValueNameError(Exception):
+    def __init__(self, cls):
         super().__init__(
-            "TimeBracket values haven't been assigned yet, run"
-            " 'assign_concrete_times' method first"
+            f"Confusing setup in class '{cls}' which inherits from "
+            "'SingleValueParameter'. Found already a defined value for"
+            " 'value_name'. SingleValueParameter instances can only "
+            "have one value!"
         )
 
 
-class OverlappingTimeBracketsError(Exception):
+class InvalidStartValueError(Exception):
     def __init__(
         self,
-        previous_end_time: typing.Optional[float] = None,
-        start_time: typing.Optional[float] = None,
-        time_bracket=None,
+        start,
+        duration,
     ):
-        message = "Found overlapping time brackets!"
-        if previous_end_time and start_time:
-            previous_end_time = round(previous_end_time / 60, 2)
-            start_time = round(previous_end_time / 60, 2)
-            message += f"\nThe end time of the previous time bracket is {previous_end_time}. The start time of the current time bracket is {start_time}."
-        if time_bracket:
-            message += f"\nThe problematic candidate is:\n{repr(time_bracket)[:50]}."
-        super().__init__(message)
+        super().__init__(
+            f"Invalid value for start = '{start}' in 'squash_in' call "
+            f"for event with duration = '{duration}'!"
+            " Start has to be equal or smaller than the events duration."
+        )
+
+
+class InvalidAverageValueStartAndEndWarning(RuntimeError):
+    def __init__(self):
+        super().__init__("Average value for range where start == end is always 0!")
+
+
+class InvalidPointError(Exception):
+    def __init__(self, point, point_count):
+        super().__init__(
+            f"Found invalid point: '{point}' with {point_count} "
+            "items! Points should have two or three items."
+        )
+
+
+class ImpossibleToSquashInError(TypeError):
+    def __init__(self, event_to_be_squashed_into, event_to_squash_in):
+        super().__init__(
+            f"Can't squash '{event_to_squash_in}' in '{event_to_be_squashed_into}'. "
+            "Does the SimultaneousEvent contain SimpleEvents or events that inherit"
+            " from SimpleEvent? For being able to squash in, the"
+            " SimultaneousEvent needs to only contain SequentialEvents or"
+            " SimultaneousEvents."
+        )
+
+
+class InvalidStartAndEndValueError(Exception):
+    def __init__(self, start, end):
+        super().__init__(
+            f"Invalid values for start and end property (start = '{start}' "
+            f"and end = '{end}')!"
+            " Value for 'end' has to be bigger than value for 'start'."
+        )
+
+
+class InvalidCutOutStartAndEndValuesError(Exception):
+    def __init__(self, start, end, simple_event, duration):
+        super().__init__(
+            f"Can't cut out SimpleEvent '{simple_event}' with "
+            f"duration '{duration}' from"
+            f" (start = {start} to end = {end})."
+        )
 
 
 class SplitUnavailableChildError(Exception):