+++ mutwo/core_utilities/prime_factors.py
@@ -17,19 +17,19 @@ class PrimeGenerator(object):
     """Generator that yields the rising series of primes starting from 2."""
 
     def __init__(self):
-        self.it = primesieve.Iterator()
+        self.iterator = primesieve.Iterator()
 
     def __next__(self) -> int:
-        return self.it.next_prime()
+        return self.iterator.next_prime()
 
     def __iter__(self) -> typing.Iterator:
         return self
 
 
-def factorise(n: int) -> list[int]:
+def factorise(number_to_factorise: int) -> list[int]:
     """factorise(integer) -> [list of factors]
 
-    :param n: The number which shall be factorised.
+    :param number_to_factorise: The number which shall be factorised.
     :return: Returns a list of the (mostly) prime factors of integer n. For negative
         integers, -1 is included as a factor. If n is 0, 1 or -1, [n] is
         returned as the only factor. Otherwise all the factors will be prime.
@@ -41,14 +41,16 @@ def factorise(n: int) -> list[int]:
     >>> factorise(55614)
     [2, 3, 13, 23, 31]
     """
-    result = []
-    for p, count in factors(n):
-        result.extend([p] * count)
-    return result
+    result_list = []
+    for prime_number, prime_number_count in factors(number_to_factorise):
+        result_list.extend([prime_number] * prime_number_count)
+    return result_list
 
 
-def factors(n: int):
-    """factors(integer) -> yield factors of integer lazily
+def factors(number: int) -> typing.Generator:
+    """Get factor generator
+
+    :param number: The number from which to yield factors.
 
     Yields tuples of (factor, count) where each factor is unique and usually
     prime, and count is an integer 1 or larger.
@@ -62,30 +64,30 @@ def factors(n: int):
     >>> list(factors(3*7*7*7*11))
     [(3, 1), (7, 3), (11, 1)]
     """
-    if n in (0, 1, -1):
-        yield (n, 1)
+    if number in (0, 1, -1):
+        yield (number, 1)
         return
-    elif n < 0:
+    elif number < 0:
         yield (-1, 1)
-        n = -n
-    assert n >= 2
-    for p in PrimeGenerator():
-        if p * p > n:
+        number = -number
+    assert number >= 2
+    for prime_number in PrimeGenerator():
+        if prime_number * prime_number > number:
             break
         count = 0
-        while n % p == 0:
+        while number % prime_number == 0:
             count += 1
-            n //= p
+            number //= prime_number
         if count:
-            yield (p, count)
-    if n != 1:
-        yield (n, 1)
+            yield (prime_number, count)
+    if number != 1:
+        yield (number, 1)
 
 
-def is_prime(n: int) -> bool:
+def is_prime(number_to_test: int) -> bool:
     """Test if number is prime or not.
 
-    :param n: The number which shall be tested.
+    :param number_to_test: The number which shall be tested.
     :return: True if number is prime and False if number isn't a Prime.
 
     (`has been copied from here
@@ -94,19 +96,19 @@ def is_prime(n: int) -> bool:
     """
 
     # Corner cases
-    if n <= 1:
+    if number_to_test <= 1:
         return False
-    if n <= 3:
+    if number_to_test <= 3:
         return True
 
     # This is checked so that we can skip
     # middle five numbers in below loop
-    if n % 2 == 0 or n % 3 == 0:
+    if number_to_test % 2 == 0 or number_to_test % 3 == 0:
         return False
 
     i = 5
-    while i * i <= n:
-        if n % i == 0 or n % (i + 2) == 0:
+    while i * i <= number_to_test:
+        if number_to_test % i == 0 or number_to_test % (i + 2) == 0:
             return False
         i = i + 6